+++ src/core/dispatcher.js
@@ -44,7 +44,6 @@
 /* global MessageChannel */
 
 const buffers = require('./buffers');
-const { options } = require('./options');
 
 const queue = buffers.queue();
 const EMPTY = buffers.EMPTY;
@@ -53,7 +52,28 @@ const SET_IMMEDIATE = 0;
 const MESSAGE_CHANNEL = 1;
 const SET_TIMEOUT = 2;
 
+const options = {
+  batchSize: 1024,
+  dispatchMethod: SET_IMMEDIATE
+};
+
 let dispatcher = createDispatcher();
+
+// Sets the values of one or more configurable options. This function takes an object for a parameter, and if any of
+// that object's property's names match one of those of the options object, its value becomes the new value of that
+// option. Properties that don't exist as options are ignored, and options that are not present in the parameter are
+// left unchanged.
+function config(opts) {
+  for (const key in options) {
+    if (opts.hasOwnProperty(key)) {
+      options[key] = opts[key];
+
+      if (key === 'dispatchMethod') {
+        setDispatcher();
+      }
+    }
+  }
+}
 
 let running = false;
 let queued  = false;
@@ -62,13 +82,13 @@ let queued  = false;
 // methods should be used.
 function getDispatchMethod() {
   switch (options.dispatchMethod) {
-    case 'MessageChannel':
+    case MESSAGE_CHANNEL:
       if (typeof MessageChannel !== 'undefined') {
         return MESSAGE_CHANNEL;
       }
       return SET_TIMEOUT;
 
-    case 'setTimeout':
+    case SET_TIMEOUT:
       return SET_TIMEOUT;
 
     default:
@@ -160,6 +180,9 @@ function dispatch(task) {
 }
 
 module.exports = {
-  setDispatcher,
-  dispatch
+  config,
+  dispatch,
+  SET_IMMEDIATE,
+  MESSAGE_CHANNEL,
+  SET_TIMEOUT
 };